requirements (1).txt
# ── Core LLM Framework ─────────────────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langgraph==0.2.28

# ── Agent Framework ────────────────────────────────────────────────────────────
crewai==0.63.6
crewai-tools==0.12.1

# ── Vector Database ────────────────────────────────────────────────────────────
chromadb==0.5.5

# ── Graph Database ─────────────────────────────────────────────────────────────
neo4j==5.24.0

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers==3.1.1
torch>=2.0.0

# ── Document Processing ────────────────────────────────────────────────────────
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
unstructured==0.15.13
markdown==3.7

# ── Settings and Validation ────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ── API Layer ─────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# ── Tokenization ──────────────────────────────────────────────────────────────
tiktoken==0.7.0

# ── Reliability ───────────────────────────────────────────────────────────────
tenacity==9.0.0

# ── Observability ─────────────────────────────────────────────────────────────
loguru==0.7.2
rich==13.8.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
